Add an API for adding hotel room

+++ iTechArt_Booking.Domain/Models/Room.cs
@@ -10,7 +10,7 @@ namespace iTechArt_Booking.Domain.Models
     {
         public long Id { get; set; }
         public string Category { get; set; }
-        public float Cost_per_day { get; set; }
+        public decimal Cost_per_day { get; set; }
         public long Number_of_beds { get; set; }
 
         public long Hotel_id { get; set; }
